backend/requirements.txt
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
social-auth-app-django==5.4.1
python-ulid==2.7.0
celery[redis]==5.4.0
flake8==7.1.0
coverage==7.6.0
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
django-elasticsearch-dsl==8.0
python-slugify[unidecode]==8.0.4
django-autoslug==1.9.9
drf-yasg==1.21.7
django-redis==5.4.0
ddt==1.7.2
tblib==3.0.0